// templates/admin/src/api/generated/types.ts
export interface components {
    schemas: {
        ProjectInput: {
            name: string;
            owner: string;
        };
        Project: {
            id: string;
            name: string;
            owner: string;
            /** @enum {string} */
            status: "active" | "paused" | "completed";
            /** Format: date */
            updatedAt: string;
        };
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}